Struct context_bind::Routine

pub struct Routine {
    pub data: usize,
    pub init: bool,
    pub id: RoutineID,
    pub state: Status,
    // some fields omitted
}

Encapsulate the state of a co-routine

This holds the entire state of a co-routine. Some fields are public to allow for eaiser inspection.

Fields
data: usize init: bool id: RoutineID state: Status
Methods

impl Routine

fn new(b: FN, id: RoutineID, stack: StackSize) -> Result<Routine, usize>

Build a new routine.

If you call this in a routine horrible things will happen. Your new routine will panic on entry, and the old routine will panic on exit.

This function will return an error if the stack created is too large. The maximum stack size your process may allocate will be listed in the Err field if that occurs.

fn exec(&mut self, data: usize) -> usize

Run the Routine.

This behaves identical to swap, except it is attached to an object.